### 2016-06-20 #30 @whedon
[REVIEW]: Backtest Graphics
**Submitting author:** @karantibrewal (<a href="http://orcid.org/0000-0003-2717-5006">Karan Tibrewal</a>)
**Repository:** <a href="https://github.com/knightsay/backtestGraphics" target ="_blank">https://github.com/knightsay/backtestGraphics</a>
**Version:** v1.0.0
