### 2016-06-20 #30 @arfon
@karantibrewal - thanks for submitting this. I have a couple of questions/comments before we can proceed with this submission:

- Please update your [`paper.md`](https://github.com/knightsay/backtestGraphics/blob/master/paper.md) metadata. It still has me listed as an author
- Isn't not obvious to me that this software has a research application. Could you please explain what areas of research this software is designed to serve?
