### 2016-06-20 79bb08d
some more changes for JOSS
