### 2016-06-20 #30 @karantibrewal
Hi Arfon,

Thank you for your email. To address your points:

1. Sorry about that. I've made the changes, and also included a paper.json
 file.

2. Although our package doesn't solve a research question directly, we have
designed it to help as a tool for others in visualization of financial
data. Recently, we presented the package to the RFinance Conference, and
received great feedback on how our package can be directly used by several
in their research. Another advantage is that the package can be modified to
support visualization of any time series data (doesn't have to be finance),
according to the user's preference.

I hope that answers your question. Please let me know if you need anything
else.

Thanks again,
Karan

On Mon, Jun 20, 2016 at 7:23 PM, Arfon Smith <notifications@github.com>
wrote:

> @karantibrewal <https://github.com/karantibrewal> - thanks for submitting
> this. I have a couple of questions/comments before we can proceed with this
> submission:
>
>    - Please update your paper.md
>    <https://github.com/knightsay/backtestGraphics/blob/master/paper.md>
>    metadata. It still has me listed as an author
>    - Isn't not obvious to me that this software has a research
>    application. Could you please explain what areas of research this software
>    is designed to serve?
>
