### 2016-06-21 f505e7f
changes to vignette:
- introduction changed
- description of data sets enhanced
- data frame display stylized according to Dave's specifications
