+++ vignettes/backtestGraphics.bib
@@ -40,3 +40,30 @@
     note = {R package version 0.12.1},
     url = {http://CRAN.R-project.org/package=shiny},
 }
+
+@book{equities,
+    author    = {R. Stafford Johnson},
+    booktitle = {Equity Markets and Portfolio Analysis},
+    title     = {Equity Markets and Portfolio Analysis},
+    publisher = {John Wiley & Sons},
+    address   = {},
+    year      = {2014}
+}
+
+@book{futures,
+    author    = {George Kleinman},
+    booktitle = {Trading Commodities and Financial Futures},
+    title     = {Trading Commodities and Financial Futures},
+    publisher = {FT Press},
+    address   = {Upper Saddle River, NJ 07458},
+    year      = {2013}
+}
+
+@ARTICLE{CDS,
+         AUTHOR= {Chen, Heidi and Dang, Yuanchu and Kane, David and Lu, Yang and Smith, Skylar and Malik, Kanishka and Zhu, Miller },
+         TITLE= {Creditr: Credit Default Swaps in R},
+         JOURNAL= {CRAN},
+         YEAR= {2015}
+}
+
+
