### 2016-06-21 #30 @arfon
OK thanks for the background @karantibrewal. It looks like the paper metadata is still a little wonky. It needs to be formatted as follows:

```
---
title: 'Fidgit: An ungodly union of GitHub and figshare'
tags:
  - example
  - tags
  - for the paper
authors:
 - name: Arfon M Smith
   orcid: 0000-0002-3957-2474
   affiliation: GitHub Inc.
date: 14 February 2016
bibliography: paper.bib
---
```
