### 2016-06-21 6a96c67
update paper.md
