### 2016-06-21 #30 @karantibrewal
Hi Arfon,

I've made the aforementioned changes.

Thanks,
Karan
