### 2016-06-22 #30 @desilinguist
Before I start my review, I have a somewhat basic question. I am wondering why the name of the author making the submission ("Karan Tibrewal") does not appear in a bunch of the documentation including the main README and the R vignette etc.  That should really be fixed.

Now on to my review:

1. The name of the license file is misspelled (`LISENCE` instead of `LICENSE`).

2. The release version says `0.1.5` and not `1.0.0`.

3. There are no installation instructions in the main README. However, there seem to be installation instructions in the `.Rmd` file that work just fine. These should also be added to the main README.

4. I looked at the documentation via `help(backtestGraphics)` and it seems to be fairly complete. However, there seems to be some weird formatting in that file at the end:

    ```
    Examples:

         ## Not run:

         backtestGraphics(data = credit.bt)
         ## End(Not run)
    ```

5. There are no performance claims made in the documentation so I couldn't verify them.

6. Since I don't really know that much about the financial domain, once I installed the package, it was hard for me to figure out how to try it based just on the README. However, once I ran `vignette('backtestGraphics')`, I got a really nice PDF that told me that there was sample data available and how I can try out `backtestGraphics` on that data.

7. I ran the first command suggested by the vignette `backtestGraphics(x = commodity)`, it launched the local shiny app but I couldn't get any plots to show up when I clicked on the `Visualize` button. However, the online shiny app linked to from the README does produce the graphs. So, perhaps something is broken locally or may be I am doing something wrong? I tried on both Safari and Chrome on my Mac with OS X 10.11.3.

8. Another thing that I noticed was that although a bunch of dependencies were installed when I first installed `backtestGraphics`, it prompted me to update `shiny` at the time of running the vignette example. It should really enforce that dependency at installation time.

9. There seem to be tests available but they don't seem to be run automatically and there is no documentation for how to run them.

10. I just noticed that `paper.md` still has some incorrect formatting at the top.

11. Although @karantibrewal says that this can be used for any time series data, the documentation is very, very specific to financial data and there's no documentation on how one would use this for non-financial data.

In conclusion, I think this is a niche shiny app that seems to be very focused on visualizing financial time series data. As it stands, I am not 100% sure that this has significant application in research but it can perhaps be a useful visualization tool for some researchers. I really think that the main README needs to have a lot more information and the statement of need really needs to be strengthened especially if the authors are claiming that the tool can visualize data outside of the financial domain.
