### 2016-06-22 #30 @karantibrewal
@desilinguist thanks a lot for the feedback. Sorry about the loose ends. We are working on it, and will address all of the aforementioned points by the end of this week.
