### 2016-06-23 #30 @pjotrp
It is important that it is clear what the scientific contribution is of this work. The authors can give some examples of financial research application, preferably with figures in the docs/vignette. It would be even better if the authors can think of a broader application in scientific research and expand their audience.
