### 2016-10-08 #30 @arfon
> @desilinguist thanks a lot for the feedback. Sorry about the loose ends. We are working on it, and will address all of the aforementioned points by the end of this week.

@karantibrewal - could we have a status update on this submission please? Are you still working on the feedback from @desilinguist?
