### 2016-10-10 #30 @karantibrewal
@desilinguist @arfon Sorry for the delay! My team has had some back and
forth with the feedback. We have decided on the following: BacktestGraphics
is not a package that explicitly forwards a topic of research, but instead,
it is a tool that researchers can use to visualize backtest data.

We think our tool could be very relevant for researchers. Financial data is
almost always presented as non-interactive graphs or tables which are
difficult to decipher. Our package gives a quick and efficient avenue to
improve on this (for examples of working papers that could leverage our
work, see here
<http://poseidon01.ssrn.com/delivery.php?ID=131074121084086071092103006092110093109023030014084091014027003126031066108084004090035056097013011097115069069031019089020121027020011005061088103084114104003080093061020081065105082020113117021025126103113002025085004089093106079081004015089095090073&EXT=pdf>
and here
<http://poseidon01.ssrn.com/delivery.php?ID=479004100026072008087021121123114090056069085080028027023073016081091004007096069105004042026057051007117120107082111114008040017035051049100079028086115107065002003044091016065102072117075006087096100113070123109003125096100117001011023126118103&EXT=pdf>
).

Please let me know if you have any other questions. We are excited to
contribute to the open source community, and would love to help out in any
way that we can.

On Sat, Oct 8, 2016 at 9:34 AM, Arfon Smith <notifications@github.com>
wrote:

> @desilinguist <https://github.com/desilinguist> thanks a lot for the
> feedback. Sorry about the loose ends. We are working on it, and will
> address all of the aforementioned points by the end of this week.
>
> @karantibrewal <https://github.com/karantibrewal> - could we have a
> status update on this submission please? Are you still working on the
> feedback from @desilinguist <https://github.com/desilinguist>?
>
