### 2016-10-20 b13e7ea
Update paper.md
