### 2016-10-20 1fdd20c
Update paper.md
