### 2016-10-20 d88bd37
Update paper.md
