### 2016-10-20 20b3e20
Update paper.md
