### 2016-10-20 9386bd3
Update paper.md
