### 2016-10-20 71967d9
Update paper.md
