+++ paper.md
@@ -18,8 +18,9 @@
     email: zijie.miller.zhu@gmail.com
     affiliation: Williams College    
 - date: 30th May 2016
-- bibliography: paper.bib
-- 
+- bibliography: paper.bib    
+
+
 ---
 
 ![](bg1.png)
