### 2016-10-21 ef593c4
Delete backtestGraphics_Intro.Rmd
