### 2016-10-21 103e3cf
Update paper.md
