+++ paper.md
@@ -1,6 +1,6 @@
 
 ---
-- title: 'Backtest Graphics'   
+- title: "Backtest Graphics"   
 - authors: 
   - name: David Kane
     email: david.kane@gmail.com
@@ -20,16 +20,15 @@
 - date: 30th May 2016
 - bibliography: paper.bib    
 
-
 ---
 
 ![](bg1.png)
 
 ###Summary
 
-Backtesting is the process of testing trading strategies on prior time horizons to measure the effectiveness of a given strategy. It helps investors understand and optimize their trading strategies [@backtest]. The **backtestGraphics** package provides an interactive **Shiny** interface to visualize backtest results for a variety of financial instruments (equities, futures , and credit defualt swaps, among others) [@shiny]. These visualizations enable users to employ their human perception to process a lot of backtest data quickly and approximately [@bostock]. 
+Backtesting is the process of testing trading strategies on prior time horizons to measure the effectiveness of a given strategy. It helps investors understand and optimize their trading strategies. The **backtestGraphics** package provides an interactive **Shiny** interface to visualize backtest results for a variety of financial instruments (equities, futures , and credit defualt swaps, among others). These visualizations enable users to employ their human perception to process a lot of backtest data quickly and approximately. 
 
-It is important to note here that **backtestGraphics** doesn't run backtests, but instead provides graphical visualization of backtest results. To illustrate this subtle distinction, consider a simple trading strategy that buys the top 10 shares of the S&P 500, and shorts the bottom 10. Now, to test how the strategy performs historically, the user may use backtesting software like **Quantopia**, or even  R packages like **backtest** [@backtest].The output from such backtesting is generally in the form of large dataframes which are difficult to interpret. As of such, the user may feed these dataframes, i.e. the holding data, to **backtestGraphics**. The package then constructs interactive **dygraph** plots, and calculates essential summary statistics which are easy to interpret and explore [@dygraphs]. 
+It is important to note here that **backtestGraphics** doesn't run backtests, but instead provides graphical visualization of backtest results. To illustrate this subtle distinction, consider a simple trading strategy that buys the top 10 shares of the S&P 500, and shorts the bottom 10. Now, to test how the strategy performs historically, the user may use backtesting software like **Quantopia**, or even  R packages like **backtest**.The output from such backtesting is generally in the form of large dataframes which are difficult to interpret. As of such, the user may feed these dataframes, i.e. the holding data, to **backtestGraphics**. The package then constructs interactive **dygraph** plots, and calculates essential summary statistics which are easy to interpret and explore. 
 
 The **Shiny** interface returned by the package consists of a sidebar panel that includes a "Summary" and a "Detail" tab. The former provides the user with summary statistics, such as average gross market value (GMV), number of instruments, cumulative and annualized profit and loss (P&L), Sharpe ratio and best and worst performing months. The "Detail" tab provides information about the best and worst performers, as well as the biggest drawdowns. 
 
@@ -43,5 +42,13 @@ Additionally, to accomodate for more complex backtests, **backtestgraphics** all
 
 In order to see **backtestGraphics** in action, please visit [here](https://backtestgraphics.shinyapps.io/backtestGraphics)!
 
-###References
+###Installation
+
+You can easily install backtestGraphics by typing:
+
+    install.packages("backtestGraphics")
+or
+
+    library(devtools)
+    install_github("knightsay/backtestGraphics")
 
