### 2016-10-21 fb8b298
Update paper.md
