+++ DESCRIPTION
@@ -16,6 +16,7 @@ Date: 2015-08-13
 Author: as.person(c(
                 "David Kane <dave.kane@gmail.com> [aut]",
                 "Ziqi Lu <ziqi.lu@williams.edu> [aut]",
+                "Karan Tibrewal <karan.tibrewal@williams.edu> [aut]",
                 "Fan Zhang <fan.zhang@williams.edu> [aut]",
                 "Miller Zijie Zhu <zijie.miller.zhu@gmail.com> [aut]"
              ))
