### 2016-10-21 6714283
Update paper.md
