### 2016-10-21 #30 @karantibrewal
@desilinguist @arfon

My co-author just explained to me that I should be more clear in my explanation. Let me try again. I've addressed @desilinguist's points below:

> The name of the license file is misspelled (LISENCE instead of LICENSE).

Fixed.

> The release version says 0.1.5 and not 1.0.0.

Fixed.

>  There are no installation instructions in the main README. However, there seem to be installation instructions in the .Rmd file that work just fine. These should also be added to the main README.

Fixed.

> Since I don't really know that much about the financial domain, once I installed the package, it was hard for me to figure out how to try it based just on the README. However, once I ran vignette('backtestGraphics'), I got a really nice PDF that told me that there was sample data available and how I can try out backtestGraphics on that data.

Added instructions to the README.

> I ran the first command suggested by the vignette backtestGraphics(x = commodity), it launched the local shiny app but I couldn't get any plots to show up when I clicked on the Visualize button. However, the online shiny app linked to from the README does produce the graphs. So, perhaps something is broken locally or may be I am doing something wrong? I tried on both Safari and Chrome on my Mac with OS X 10.11.3.

Apologies! Do you think you can provide me with more details on this? I've tried the package on several computers and it seems to be working fine. If you can give me more information/ screenshots, I can look into this further.

> I just noticed that paper.md still has some incorrect formatting at the top.

Fixed.

> Although @karantibrewal says that this can be used for any time series data, the documentation is very, very specific to financial data and there's no documentation on how one would use this for non-financial data.

Apologies for the miscommunication! This is a package for financial data, and specifically, for backtest results only.

> I think this is a niche shiny app that seems to be very focused on visualizing financial time series data. As it stands, I am not 100% sure that this has significant application in research but it can perhaps be a useful visualization tool for some researchers.

We agree! However, we believe that it is a fairly large niche. For example, we found several working papers on SSRN that could benefit from our package. Most of these papers present their data as tables or static graphics. Our package aims to enable these researchers to provide informative and interactive graphics, without going through the trouble of creating them themselves.

>  I really think that the main README needs to have a lot more information and the statement of need really needs to be strengthened especially if the authors are claiming that the tool can visualize data outside of the financial domain.

Our package only visualizes financial, and more specifically, backtest data. I have updated the README to make this more explicit.

@desilinguist @arfonsmith Hope this was helpful. Looking forward to any other feedback you may have!
