### 2016-10-23 #30 @arfon
@desilinguist - do you think you could take another look at this submission? @karantibrewal has responded to much of your feedback.
