### 2016-10-24 #30 @desilinguist
@karantibrewal, thanks for making all the changes. I think they help make it clear exactly what the package does and the exact scenarios under which it can he useful to researchers studying financial data. One suggestion, it would be nice to explain in the README that to get more detailed documentation and background, users should run `vignette("backtestGraphics")` for people who may be new to R.

More importantly, I am still not able to get the local shiny app to actually show me any plots. I basically followed the `commodity` example in the README and ran the two commands in RStudio and got an error after hitting the "Visualize" button. I am attaching the screenshot here.
