### 2016-10-25 #30 @desilinguist
I've been swamped at work and so haven't had a chance. I'll try and take a
look tomorrow. Thanks for the reminder :)
On Sun, Oct 23, 2016 at 1:56 PM Arfon Smith <notifications@github.com>
wrote:

> @desilinguist <https://github.com/desilinguist> - do you think you could
> take another look at this submission? @karantibrewal
> <https://github.com/karantibrewal> has responded to much of your feedback.
>
