### 2016-10-25 #30 @karantibrewal
@desilinguist thank you for your feedback. Ill go over the issues you
outlined, and revert by begining of next week.

Appreciate the help!
