### 2016-11-04 #30 @desilinguist
I removed the package and re-installed it and then ran `backtestGraphics(x = commodity)` again in RStudio. But it still doesn't work for me. No plots and same error.
