### 2016-11-04 #30 @karantibrewal
@desilinguist thanks for your prompt response. aah that’s so weird. just to
be sure, are you using a mac book?
