### 2016-11-04 #30 @desilinguist
Yes, I am using a MacBook Pro with macOS 10.12.1 with R version 3.3.0  and RStudio version 0.99.903.
