### 2016-11-18 #30 @karantibrewal
Hi @desilinguist, seems to be a dependency issue. We'll fix it for the package, but can you please confirm before that?

devtools::install_version("shiny", version="0.13.2",repos = "http://cran.us.r-project.org")
install.packages("backtestGraphics")
backtestGraphics(x = commodity)

This should work. Please let me know.
