### 2016-12-01 #30 @karantibrewal
@desilinguist: Hey Nitin, this is just a gentle follow up. If you're back,
can you please confirm the above?
