### 2016-12-02 #30 @karantibrewal
@desilinguist seems to still be a dependency issue. Can you please do the
following for me:

ip <- as.data.frame(installed.packages()[,c(1,3:4)])
ip <- ip[is.na(ip$Priority),1:2,drop=FALSE]
write.csv(ip, row.names=FALSE, file = 'package_versions.csv')

This should create a csv file for your R package/dependency versions. If
you could send that across to me, I should be able to find the issue.

Thanks
