### 2017-01-09 #30 @desilinguist
Thanks @arfon!

@karantibrewal, sorry for the delay. Here are my package versions:

| Package          	| Version   	|
|------------------	|-----------	|
| assertthat       	| 0.1       	|
| backtestGraphics 	| 0.1.6     	|
| BH               	| 1.60.0-2  	|
| colorspace       	| 1.2-7     	|
| curl             	| 2.3       	|
| DBI              	| 0.5-1     	|
| devtools         	| 1.12.0    	|
| dichromat        	| 2.0-0     	|
| digest           	| 0.6.10    	|
| dplyr            	| 0.5.0     	|
| dygraphs         	| 1.1.1.2   	|
| git2r            	| 0.16.0    	|
| htmltools        	| 0.3.5     	|
| htmlwidgets      	| 0.7       	|
| httpuv           	| 1.3.3     	|
| httr             	| 1.2.1     	|
| jsonlite         	| 1.1       	|
| labeling         	| 0.3       	|
| lazyeval         	| 0.2.0     	|
| magrittr         	| 1.5       	|
| manipulate       	| 0.98.1103 	|
| memoise          	| 1.0.0     	|
| mime             	| 0.5       	|
| munsell          	| 0.4.3     	|
| openssl          	| 0.9.5     	|
| plyr             	| 1.8.4     	|
| R6               	| 2.2.0     	|
| RColorBrewer     	| 1.1-2     	|
| Rcpp             	| 0.12.7    	|
| rstudio          	| 0.98.1103 	|
| rstudioapi       	| 0.6       	|
| scales           	| 0.4.0     	|
| shiny            	| 0.13.2    	|
| sourcetools      	| 0.1.5     	|
| tibble           	| 1.2       	|
| whisker          	| 0.3-2     	|
| withr            	| 1.0.2     	|
| xtable           	| 1.8-2     	|
| xts              	| 0.9-7     	|
| yaml             	| 2.1.13    	|
| zoo              	| 1.7-13    	|
