### 2017-02-16 d83af0d
prev. render table alignment deprecated -- fixed accordingly.
