### 2017-02-25 #30 @karantibrewal
Hi @desilinguist,
I've tried recreating your bug in the past month on a bunch of computers but with little success -- however, I'm eager to work with you to fix this, and I appreciate your patience.

One possible reason that the package might be failing is if "dygraph" isn't set up properly on your computer. The following code snippet will test that:

--------------

library(backtestGraphics)
library(dygraphs)
library(dplyr)
library(xts)

sum_data <- function(x, sector.selected = NULL, capital.num = NULL){

  sector <- nmv <- pnl <- contract <- gmv <- NULL

  if(!is.null(sector.selected)) {
    x <- filter(x, sector == sector.selected)
  }

  if(dim(x)[1] == 0){
    stop("There does not exist any data that satisfies your selection.
         Please select a new combination of Strategy, Portfolio and Instrument.")
  }

  x <- group_by(x, date)

  x.sum <- summarise(x,
                     nmv = sum(nmv, na.rm = TRUE),
                     pnl = sum(pnl, na.rm = TRUE),
                     contract = sum(contract, na.rm=TRUE),
                     gmv = sum(gmv, na.rm = TRUE))
  x.sum <- tbl_df(x.sum) %>%
    arrange(date) %>%
    filter(gmv != 0)

  if(is.null(capital.num)){
    x.sum <- mutate(x.sum, ret = pnl / gmv, cumpnl = cumsum(pnl))
  } else {
    x.sum <- mutate(x.sum, ret = pnl / capital.num, cumpnl = cumsum(pnl))
  }

  x.sum <- mutate(x.sum, name = " ")

  if(!is.null(sector.selected)){
    x.sum[[1,8]] <- sector.selected
  } else {
    x.sum[[1,8]] <- "the Whole Portfolio"
  }

  return(x.sum)

}

slice_data <- function(x, input, capital.num){

  name <- id <- portfolio <- gmv <- nmv <- contract <- x.name <- strategy <- NULL
  substrategy <- capital.num <- pnl <- sector <- NULL

  x.name <- unique(select(x, name, id, sector, portfolio, strategy, substrategy))

  f <- list()
  f$x.temp <- x %>%
    group_by(name, id, date, sector, portfolio) %>%
    summarise(gmv = sum(gmv, na.rm = TRUE),
              nmv = sum(nmv, na.rm = TRUE),
              pnl = sum(pnl, na.rm = TRUE),
              contract = sum(abs(contract), na.rm = TRUE)) %>%
    ungroup()
  f$x.temp <- f$x.temp %>%
    group_by(name, id, date, sector) %>%
    summarise(gmv = sum(gmv, na.rm = TRUE),
              nmv = sum(nmv, na.rm = TRUE),
              pnl = sum(pnl, na.rm = TRUE),
              contract = sum(abs(contract), na.rm = TRUE)) %>%
    ungroup()
  f$x <- sum_data(f$x.temp,
                  capital.num = capital.num)

  f$instrument <- length(unique(f$x.temp[["id"]]))
  return(f)
}

cumpnl_plot <- function(x) {

  date <- name <- cumpnl <- NULL

  title.var <- paste("Cumulative P&L for", x$name[1])
  x <- xts(x$cumpnl, order.by = x$date)
  dygraph(x, main = title.var) %>%
    dySeries(label = "Cumulative P&L") %>%
    dyRangeSelector() %>%
    dyLegend(labelsSeparateLines = TRUE) %>%
    dyOptions(labelsKMB = TRUE,
              useDataTimezone = TRUE,
              fillGraph = TRUE)
}

data <- slice_data(commodity, NULL, NULL)
p1 <- cumpnl_plot(x = data$x)
p1

-----------------

Please run the above on your R studio console. You should hopefully see a plot like:

Please let me know what you see. Thanks.
