### 2017-03-20 #30 @desilinguist
When I run your code snippet, I get no errors but the graph is blank as shown below.
