### 2017-03-20 #30 @karantibrewal
Thanks @desilinguist. I believe there is a problem with the dygraphs package on your computer. I tried running the same commands with your version, but the graph was not blank. Could you try the following:

(1) Try reinstalling dygraphs, and running the previous code.
That is,
```
remove.packages("dygraphs")
install.packages("dygraphs")
<previous code>
```

(2) If you still see a blank plot, try a simpler plot:
```
library(dygraphs)
library(xts)
df <- data.frame(x = 1:10, y = as.Date(seq(as.Date("2017/1/1"), as.Date("2017/1/10"), "days")))
x <- xts(df$x, order.by = df$y)
dygraph(x)
```

Please let me know what you see.
