### 2017-03-20 #30 @desilinguist
I still see a blank graph after removing and re-installing `dygraphs`, for both snippets.
