### 2017-03-20 #30 @karantibrewal
Thanks for your prompt reply @desilinguist! This is definitely an issue with the Dygraphs package. I would suggest you to delete and reinstall R, RStudio, and try running the simple plot again:

```
install.packages(c("dygraphs", "xts"))
library(dygraphs)
library(xts)
df <- data.frame(x = 1:10, y = as.Date(seq(as.Date("2017/1/1"), as.Date("2017/1/10"), "days")))
x <- xts(df$x, order.by = df$y)
dygraph(x)
```
