### 2017-03-20 #30 @desilinguist
Even if I run on the terminal in `R`, I still get a blank graph.

Perhaps this should be assigned to another reviewer. I don't think I want to delete and reinstall everything.
