### 2017-03-20 #30 @karantibrewal
Sorry for the inconvenience @desilinguist. I'm fairly certain that this is a problem in your local machine since simple dygraph plots don't show up. However, I understand your predicament in having to delete and reinstall everything. Thanks for your help in the process.
