### 2017-04-03 #30 @karantibrewal
Thanks again, @desilinguist!

Hi @arfon! I believe the issue was with dygraphs (a graphing package in R).
Otherwise, the package builds successfully. I could provide any further
details required by your team.

On Mon, Apr 3, 2017 at 8:57 AM, Arfon Smith <notifications@github.com>
wrote:

> @arfon <https://github.com/arfon> Perhaps this should be assigned to
> another reviewer? I don't really want to delete and reinstall everything
> just for this one review.
>
> 👍 ok no problem @desilinguist <https://github.com/desilinguist> - thanks
> for all of your help thus far.
>
> @karantibrewal <https://github.com/karantibrewal> - do you have any
> suggestions for alternative reviewers for this?
>
