### 2017-04-05 #30 @arfon
> Hi @arfon! I believe the issue was with dygraphs (a graphing package in R). Otherwise, the package builds successfully. I could provide any further details required by your team.

OK thanks. It's not clear to me whether you're suggesting here that we _don't_ need a followup review of this? Could you advise @karantibrewal & @desilinguist?
