### 2017-04-10 #30 @karantibrewal
Sorry @arfon, I think I might have misunderstood you!

I agree that we require another review. Do you mean suggestions for people who could be reviewers? I don't have anyone at the top of my mind, but I could ask around in my network if people are willing. Are there any guidelines for selecting these reviewers?
