### 2017-04-14 #30 @karantibrewal
Thanks Arfron. I'll get back to you on this.

On Thu, Apr 13, 2017 at 8:36 PM Arfon Smith <notifications@github.com>
wrote:

> Do you mean suggestions for people who could be reviewers? I don't have
> anyone at the top of my mind, but I could ask around in my network if
> people are willing.
>
> Yes, that's what I mean.
>
> Are there any guidelines for selecting these reviewers (I see the
> guidelines for reviewing; what I mean is that is there any minimum
> qualification/requirements to sign up as a reviewer)?
>
> Not really, other than that they should be experienced in the topic of the
> submission and the technologies used.
>
