### 2017-07-10 #30 @arfon
@openjournals/ropensci - any chance you could help me find a second reviewer here?
