### 2017-07-10 #30 @noamross
https://github.com/joshuaulrich is an experienced R developer in financial time-series and has volunteered to review for us before.
