### 2017-07-10 #30 @arfon
@joshuaulrich - would you be interested in reviewing this submission for JOSS?
