### 2017-07-10 #30 @joshuaulrich
@arfon I'd be happy to review. I remember this package from [R/Finance 2016](http://www.rinfinance.com/RinFinance2016/agenda/).
