### 2017-07-10 #30 @arfon
> @arfon I'd be happy to review. I remember this package from R/Finance 2016.

@joshuaulrich - many thanks! Please update the checklist above and giving feedback in this issue. The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines
