### 2017-08-28 #30 @joshuaulrich
@karantibrewal Very strange that the GitHub version is *behind* the CRAN version.  It appears the CRAN version has changes to `R/backtestGraphics.R` that are not in the GitHub repository.
