### 2017-08-28 knightsay/backtestGraphics#1 @joshuaulrich
Unescaped "&" prevent vignette build
The [backtestGraphics.Rnw](https://github.com/knightsay/backtestGraphics/blob/master/vignettes/backtestGraphics.Rnw) vignette does not build because there are several unescaped `&`, which causes `texi2dvi` to fail.  Therefore I cannot build/install the package without `--no-bulid-vignettes`.

```
> /usr/bin/R CMD build backtestGraphics
* checking for file 'backtestGraphics/DESCRIPTION' ... OK
* preparing 'backtestGraphics':
* creating vignettes ... ERROR

Attaching package: ‘dplyr’

The following objects are masked from ‘package:stats’:

    filter, lag

The following objects are masked from ‘package:base’:

    intersect, setdiff, setequal, union

Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet,  :
  Running 'texi2dvi' on 'backtestGraphics.tex' failed.
LaTeX errors:
! Misplaced alignment tab character &.
l.41 ...tegy that buys the top 10 shares of the S&
                                                  P 500, and shorts the bott...
I can't figure out why you would want to use a tab mark
here. If you just want an ampersand, the remedy is
! Misplaced alignment tab character &.
l.44 ...ulative and annualized profit and loss (P&
                                                  L), Sharpe ratio and best ...
I can't figure out why you would want to use a tab mark
here. If you just want an ampersand, the remedy is
! Misplaced alignment tab character &.
l.47 ... plots for cumulative and point-in-time P&
                                                  L, NMV, GMV and number of ...
I can't figure out why you would want to use a tab mark
here. If you just want an ampersand, the remedy is
! Missing } inserted.
<inserted text>
                }
l.87 \end{itemize}

Calls: <Anonymous> -> texi2pdf -> texi2dvi
Execution halted
```

See https://github.com/openjournals/joss-reviews/issues/30.
