### 2017-09-10 #30 @joshuaulrich
@karantibrewal I'm unable to continue reviewing until you tell me which version of the package I should look at (CRAN or GitHub).  If the GitHub version is the latest, please fix the issue that prevents the package from building.
