### 2017-09-12 aa49177
fixed vignette issues -- pkg builds properly now.
