+++ vignettes/backtestGraphics.bib
@@ -12,7 +12,7 @@
     title = {Time Series Visualization with Cubism.js},
     author = {Mike Bostock},
     year = {2012},
-    url = {https://bost.ocks.org/mike/cubism/intro/#0},
+    url = {https://bost.ocks.org/mike/cubism/intro/\#0},
 }
 
 
@@ -45,7 +45,7 @@
     author    = {R. Stafford Johnson},
     booktitle = {Equity Markets and Portfolio Analysis},
     title     = {Equity Markets and Portfolio Analysis},
-    publisher = {John Wiley & Sons},
+    publisher = {John Wiley \& Sons},
     address   = {},
     year      = {2014}
 }
