### 2017-09-12 knightsay/backtestGraphics#1 @karantibrewal
Resolved, thanks for flagging.
