### 2017-09-12 #30 @karantibrewal
@joshuaulrich sorry for delay. GitHub version is latest -- issues have been fixed now. Thanks for flagging.
