### 2017-10-04 #30 @arfon
@joshuaulrich - over to you I think for another look :-)
