### 2017-11-25 #30 @joshuaulrich
Sorry for the delay @arfon.  Started a new job on 10/9 and added child number 3 on 11/1.  It's been a busy couple of months.

@karantibrewal Here are some minor comments on the PDF:

- p2 ¶1 Unnecessary space between "futures" and comma (", futures , and")
- p2 ¶2 s/Quantopia/Quantopian/
- p2 ¶4 The abbreviation NMV is not previously defined
- p2 ¶4 Reference the figure number instead of "Below is a screen shot of the plots". E.g., "Figure 2 contains a screenshot of the plots"
- p2 ¶5 s/backtestgraphics/backtestGraphics/

And some comments on the package:

The GitHub version (0.1.5) is still behind the CRAN version (0.1.6). Ideally, the Version and Date fields in the DESCRIPTION file should be incremented.

I initially thought the Shiny interface was not working because no data are displayed when it opens.  It would be helpful if the initial Shiny page contained instructions or a note that you have to click the [Visualize] button before you will see any results.

The "Strategies" drop down input contains the 3 strategies in `commodity$strategy`, but also contains 7 other strategies (e.g. Strategy 1.1).  It's not clear what those are. You could use the "optgroup" functionality of [selectInput](https://shiny.rstudio.com/reference/shiny/latest/selectInput.html) to label the portfolio-level sub-strategies.

The previous comment also applies to the "Instruments" drop down input.

You should try to use [synchronization](https://rstudio.github.io/dygraphs/gallery-synchronization.html) to link the zooming of the plots.

It's very difficult to interpret the P&L alongside the market value for any aggregate because they can simultaneously move in opposite directions if a product is added/removed from the aggregate. For example, look at the livestock aggregate for strategy 1 and portfolio 1 on in 2005-11-22.  The screenshot below shows a positive P&L of ~13.5k while the gross market value falls by ~3.5m.  This is because substrategy 1.1 has observations on 2005-11-21 but not on 2005-11-22 and the market value of substrategy 1.1 is not carried forward to days it does not have P&L.

![screenshot from 2017-11-25 09-56-24](https://user-images.githubusercontent.com/1023193/33232198-4ef5b984-d1c7-11e7-8071-d7a46d007c1c.png)

The GitHub repository does not contain instructions for how to contribute to the project or seek help.  You can look at the [quantmod contributing guide](https://github.com/joshuaulrich/quantmod/blob/master/CONTRIBUTING.md) as an example.  It contains several references to other contributing guides.  You might also consider adding an issue and/or pull request template.  I have an [issue template for quantmod](https://github.com/joshuaulrich/quantmod/blob/master/.github/issue_template.md).
