### 2017-12-06 #30 @arfon
@karantibrewal - :wave: - did you get a chance to review @joshuaulrich's feedback yet?
