### 2018-01-15 f04ebf0
Turning in comments for JOSS review [1] Fixed vignette [2] Fixed Description file (version + date) [3] Added helpText for visualize button
