+++ DESCRIPTION
@@ -11,8 +11,8 @@ Description: Creates an interactive graphics
              portfolio holdings are shown in the left panel, and interactive 
              plots of profit and loss (P\&L), net market value (NMV) and 
              gross market value (GMV) are displayed in the right panel. 
-Version: 0.1.5
-Date: 2015-08-13
+Version: 0.1.6
+Date: 2018-01-14
 Author: as.person(c(
                 "David Kane <dave.kane@gmail.com> [aut]",
                 "Ziqi Lu <ziqi.lu@williams.edu> [aut]",
