### 2018-01-15 3e4b3bd
optGroups for selectizeInput of strategy/substrategy, instrument (sector/id) and accompanying functions to create groupings
