### 2018-01-15 96920b0
more fixes for JOSS - synchronize to coordinate plot zoom in - vignette built and reloaded
