### 2018-02-05 #30 @arfon
Ping @karantibrewal - are you still interested in pursuing this publication with JOSS?
