### 2018-02-05 #30 @karantibrewal
Oops - sorry about this. Thought I had responded.

@joshuaulrich comments have been fixed now. Details are below:

> p2 ¶1 Unnecessary space between "futures" and comma (", futures , and")
> p2 ¶2 s/Quantopia/Quantopian/
> p2 ¶4 The abbreviation NMV is not previously defined
> p2 ¶4 Reference the figure number instead of "Below is a screen shot of the plots". E.g., "Figure 2 contains a screenshot of the plots"
> p2 ¶5 s/backtestgraphics/backtestGraphics/

[Fixed. Although not quite sure what the last point means. I think we need to use the name of the package instead of saying "our package". I changed that.]

> The GitHub version (0.1.5) is still behind the CRAN version (0.1.6). Ideally, the Version and Date fields in the DESCRIPTION file should be incremented.

Fixed

> I initially thought the Shiny interface was not working because no data are displayed when it opens. It would be helpful if the initial Shiny page contained instructions or a note that you have to click the [Visualize] button before you will see any results.

Fixed.

> The "Strategies" drop down input contains the 3 strategies in commodity$strategy, but also contains 7 other strategies (e.g. Strategy 1.1). It's not clear what those are. You could use the "optgroup" functionality of selectInput to label the portfolio-level sub-strategies.
>
> The previous comment also applies to the "Instruments" drop down input.

Drop downs look much more informative now - thanks for this one!

> You  should try to use synchronization to link the zooming of the plots.

Fixed.

> It's very difficult to interpret the P&L alongside the market value for any aggregate because they can simultaneously move in opposite directions if a product is added/removed from the aggregate. For example, look at the livestock aggregate for strategy 1 and portfolio 1 on in 2005-11-22. The screenshot below shows a positive P&L of ~13.5k while the gross market value falls by ~3.5m. This is because substrategy 1.1 has observations on 2005-11-21 but not on 2005-11-22 and the market value of substrategy 1.1 is not carried forward to days it does not have P&L.

This scenario can very well happen, and as I will explain now, it is not unreasonable.

Market value is a stock and P&L is a flow.

A stock is something that you measure at a specific moment in time. So, the market value of our cattle futures position is, roughly, number of contracts at a point in time multiplied by price per contract. Since the data we provide at a daily frequency, we only get a market value once per day.

P&L, on the other hand, is a flow - so it takes into account each trade you make. So for example, say you own 3 futures contracts today worth $ 1 mil. Tomorrow, you make day trades where you lose $50k. On close of market, you might still own 3 contracts worth $1 mil. In this case, your PL will be -50K but there will be no change to market value.
