### 2018-03-17 #30 @joshuaulrich
@afron, @karantibrewal,

Will review this weekend. I did notice that [backtestGraphics was removed from CRAN](https://cran.r-project.org/web/packages/backtestGraphics/index.html) since late January because `R CMD check` problems were not addressed.  I ran `R CMD check` locally and on the default platforms for `rhub::check_for_cran()`, but I did not see any obvious issues. @karantibrewal, what issue(s) did CRAN ask to be fixed?
