### 2018-03-17 knightsay/backtestGraphics#2 @joshuaulrich
Help file example does not run
`?backtestGraphics` has the following example that does not run:
```r
backtestGraphics(data = credit.bt)
```
It does not run for 2 reasons: 1) `data` is not an argument to `backtestGraphics()`, and 2) the `credit.bt` object does not exist.  The example could be changed to `backtestGraphics(commodity)`, for example.

You could also remove `\dontrun`, and add `if (interactive())` like the shiny package examples.  This would allow a user to run the example by running the command `example("backtestGraphics")`.  So the example would be:
```r
if (interactive()) {
  backtestGraphics(commodity)
}
```
See https://github.com/openjournals/joss-reviews/issues/30.
