### 2018-03-17 #30 @davidkane9
D'uh! Missed that notice because we failed to maintain a current maintainer
e-mail address. So, not sure what CRAN wanted fixed.

Karan: Please make yourself the maintainer and resubmit to CRAN.

On Sat, Mar 17, 2018 at 12:42 PM, Joshua Ulrich <notifications@github.com>
wrote:

> @afron <https://github.com/afron>, @karantibrewal
> <https://github.com/karantibrewal>,
>
> Will review this weekend. I did notice that backtestGraphics was removed
> from CRAN
> <https://cran.r-project.org/web/packages/backtestGraphics/index.html>
> since late January because R CMD check problems were not addressed. I ran R
> CMD check locally and on the default platforms for rhub::check_for_cran(),
> but I did not see any obvious issues. @karantibrewal
> <https://github.com/karantibrewal>, what issue(s) did CRAN ask to be
> fixed?
>
