### 2018-03-17 #30 @joshuaulrich
Oops, I did not respond to questions about my initial comments.

> Although not quite sure what [p2 ¶5 s/backtestgraphics/backtestGraphics/] means.

The "s/foo/bar/" means substitute "foo" with "bar".  It's a command you can do in many text programming languages (e.g. sed, awk) and is shorthand many programmers use.  Sorry for the confusion!  In this case, it means you need to change "backtestgraphics" with "backtestGraphics".  The specific text is, "...more complex backtests, \pkg{backtestgraphics} allows the user...".

----

> This scenario can very well happen, and as I will explain now, it is not unreasonable.

Right, I understand why it happens and I agree that the results are reasonable.  My point is that how they are displayed makes it hard to interpret the aggregate results.  I don't necessarily think this is a blocker to publication, but I do think it will cause confusion.  I would try plotting aggregates as stacked bar charts of their components, and/or annotating the new bar with the current instrument count whenever it is not the same as the prior bar's instrument count.

----

New comments:
* In the summary table, "Allocated Capital" should be changed to "Allocated Capital (AC)". That will provide context for the "Annualized Return on AC (%)" and "Annualized Volatility on AC (%)" rows.
* Still need a contributing guide.
* With `paper.md`:
  * ¶1 has an unnecessary space between "futures" and comma (", futures , and")
  * ¶1 "default" is incorrectly spelled "defualt"
  * ¶5 accommodate"" is incorrectly spelled "accomodate"

Conclusion: My opinion is that this is ready for publication after the minor copy edits are done, the contributing guide is added to the GitHub repo, the example runs, and the package is back on CRAN.

----

Things to address for CRAN:
* Need to add `.bib`, `.json`, and `paper.md` to `.Rbuildignore`.  Or put them in a new directory and add the directory name to `.Rbuildignore`.  Whatever you choose, just make sure they're not in the package tarball you submit to CRAN (`R CMD check --as-cran` will warn you about them).
* Add `LICENSE` to `DESCRIPTION` (again, `R CMD check --as-cran` will warn about this).
