### 2018-03-17 knightsay/backtestGraphics#3 @joshuaulrich
Add contributing guide
You can look at the [quantmod contributing guide](https://github.com/joshuaulrich/quantmod/blob/master/CONTRIBUTING.md) as an example.  It contains several references to other contributing guides.  You might also consider adding an issue and/or pull request template.  I have an [issue template for quantmod](https://github.com/joshuaulrich/quantmod/blob/master/.github/issue_template.md).

See https://github.com/openjournals/joss-reviews/issues/30.
