### 2018-03-20 7a3d680
Create CONTRIBUTING.md
Contributing guidelines for JOSS
