### 2018-03-20 092a926
(1) Fixed typos in Paper.md (2) "Allocated Capital" in Summary table changed to read "Allocated Capital (AC)" for ease of reference.
