### 2018-04-11 #30 @karantibrewal
Thanks for your comments @joshuaulrich. They have been fixed now.

However, I can't get the CRAN issue fixed for the life of me. It seems like a file path issue. The error message specifically is:

`Messages:
sh: /usr/local/opt/texinfo/bin/texi2dvi: No such file or directory
Calls: <Anonymous> -> texi2pdf -> texi2dvi
Execution halted`

Do you have any idea of how to fix this? Any advice would be much appreciated!
