### 2018-04-14 #30 @joshuaulrich
@karantibrewal I can't see any of your fixes.  In fact, I don't see any commits in the repo since 3/19.  The error looks like you don't have `texi2dvi` installed.  On Ubuntu, that's in the `texinfo` package.
