### 2018-04-15 #30 @karantibrewal
@joshuaulrich Sorry, what do you mean? All your comments from your previous post have been turned in (except for CRAN issue). These fixes were part of the 3/19 commit.
