### 2018-04-15 #30 @joshuaulrich
@karantibrewal I was confused because you said the issues were fixed, but the last activity was a month ago; and the issues I opened for them had not been closed.  It's good practice to put "Fixes #XYZ" in the commit that fixes an issue, so you have a cross-reference of the commit and issue.  Also, GitHub (and GitLab, Bitbucket, etc) will automatically close issue #XYZ for you.

And `DESCRIPTION` is still in your `.Rbuildignore`.  I commented about 2 weeks ago that this would mean the package cannot be installed.  In fact, `R CMD build backtestGraphics` fails because the package can't be installed to build the vignettes.
