### 2018-11-02 #30 @arfon
After numerous attempts to contact the reviewer, I'm assuming @karantibrewal is no longer interested in pursuing this submission in JOSS.

@joshuaulrich - thank you very much for your time, and I'm sorry this didn't make it through review.
