### 2018-11-02 #30 @joshuaulrich
No need to apologise. Let me know if there are other finance or time series
reviews I can help with.
